requirements.txt
fastapi
uvicorn
vertexai
python-dotenv
mangum
openai
langchain==0.0.123
pinecone-client
pathlib